+++ doc-delivery/src/main/java/com/wd/cloud/docdelivery/service/FrontService.java
@@ -82,14 +82,14 @@ public interface FrontService {
     Page<GiveRecordDTO> myGiveRecords(String giverName, List<Integer> status, Pageable pageable);
 
 
-    Page<HelpRecordDTO> getHelpRecords(List<Integer> channel, List<Integer> status, String email, String keyword, Boolean isDifficult, String orgFlag, Pageable pageable);
+    Page<HelpRecordDTO> getHelpRecords(List<Long> channel, List<Integer> status, String email, String keyword, Boolean isDifficult, String orgFlag, Pageable pageable);
 
     /**
      * 获取待应助的求助记录
      *
      * @return
      */
-    Page<HelpRecordDTO> getWaitHelpRecords(List<Integer> channel, Boolean isDifficult, String orgFlag, Pageable pageable);
+    Page<HelpRecordDTO> getWaitHelpRecords(List<Long> channel, Boolean isDifficult, String orgFlag, Pageable pageable);
 
     /**
      * 求助完成列表
@@ -98,7 +98,7 @@ public interface FrontService {
      * @param pageable
      * @return
      */
-    Page<HelpRecordDTO> getFinishHelpRecords(List<Integer> channel, String orgFlag, Pageable pageable);
+    Page<HelpRecordDTO> getFinishHelpRecords(List<Long> channel, String orgFlag, Pageable pageable);
 
     /**
      * 求助成功列表
@@ -107,7 +107,7 @@ public interface FrontService {
      * @param pageable
      * @return
      */
-    Page<HelpRecordDTO> getSuccessHelpRecords(List<Integer> helpChannel, String orgFlag, Pageable pageable);
+    Page<HelpRecordDTO> getSuccessHelpRecords(List<Long> helpChannel, String orgFlag, Pageable pageable);
 
     /**
      * 疑难文献（无结果，求助失败）列表
@@ -116,7 +116,7 @@ public interface FrontService {
      * @param pageable
      * @return
      */
-    Page<HelpRecordDTO> getFailedHelpRecords(List<Integer> helpChannel, List<Integer> status, String orgFlag, Pageable pageable);
+    Page<HelpRecordDTO> getFailedHelpRecords(List<Long> helpChannel, List<Integer> status, String orgFlag, Pageable pageable);
 
     DocFile getReusingFile(Long literatureId);
